feat(private-rpc): zks_getAllAccountBalances
zks_getAllAccountBalances.
