+++ packages/private-rpc/src/rpc/methods/index.ts
@@ -1,0 +1,3 @@
+export * from './eth_call';
+export * from './zks_sendRawTransaction';
+export { extractSelector } from '@/rpc/methods/utils';
