+++ packages/private-rpc/src/rpc/methods/utils.ts
@@ -1,0 +1,5 @@
+import { Hex } from 'viem';
+
+export function extractSelector(calldata: Hex): Hex {
+  return calldata.substring(0, 10) as Hex;
+}
